pyproject.toml
[project]
name = "startpage"
version = "0.1.0"
description = "Generate updates to my new startpage in Notion"
authors = [
    {name = "Dmitrii Falko",email = "dmitriy.falko@gmail.com"}
]
requires-python = ">=3.10,<4.0"
dependencies = [
    "ipykernel (>=7.1.0,<8.0.0)",
    "python-weather (>=2.1.1,<3.0.0)",
    "notion-client (>=2.7.0,<3.0.0) ; python_version >= \"3.10\" and python_version < \"4\"",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "aiohttp (>=3.9.0,<4.0.0)",
    "caldav (>=2.2.3,<3.0.0)"
]

[tool.poetry]
packages = [{include = "startpage", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.12.0,<26.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "flake8-black (>=0.4.0,<0.5.0)",
    "flake8-bugbear (>=25.11.29,<26.0.0)",
    "flake8-bandit (>=4.1.1,<5.0.0)",
    "flake8-commas (>=4.0.0,<5.0.0)",
    "flake8-builtins (>=3.1.0,<4.0.0)",
    "flake8-cognitive-complexity (>=0.1.0,<0.2.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "isort (>=7.0.0,<8.0.0)"
]
